feat: complete project structure refactoring following Python best practices
- Renamed main_refactored.py → main.py (primary entry point)
- Renamed thunder_fighter/game_refactored.py → thunder_fighter/game.py (core game class)
- Renamed thunder_fighter/graphics/ui_manager_refactored.py → thunder_fighter/graphics/ui_manager.py
- Renamed thunder_fighter/graphics/ui_components/ → thunder_fighter/graphics/ui/ (cleaner naming)
- Deleted outdated demo files and examples directory
- Updated all import statements and test references
- Added backwards compatibility methods to RefactoredGame class
- Fixed event system processing for immediate level up handling
- Fixed final boss score bonus calculation (1000 vs 500)
- All 226 tests passing

Project now follows Python naming conventions and has cleaner structure.
