+++ thunder_fighter/game.py
@@ -751,6 +751,8 @@ class RefactoredGame:
                 GameEventType.LEVEL_UP,
                 {'old_level': self.game_level, 'new_level': self.game_level + 1}
             ))
+            # Process events immediately to ensure level up is handled
+            self.event_system.process_events()
         
         # Check game over condition
         if self.player.health <= 0 and not self.game_won:
@@ -929,4 +931,89 @@ class RefactoredGame:
         
         logger.info("Game loop ended")
         pygame.quit()
-        sys.exit() 
+        sys.exit()
+    
+    # Backwards compatibility methods for testing
+    def handle_boss_defeated(self):
+        """
+        Backwards compatibility method for tests.
+        Simulates boss defeat handling.
+        """
+        if hasattr(self, '_boss_defeat_processed'):
+            return  # Prevent duplicate processing
+        
+        self._boss_defeat_processed = True
+        self._boss_defeat_reset_time = time.time() + 1.0  # Reset after 1 second
+        
+        # Get boss level before cleanup
+        boss_level = getattr(self.boss, 'level', 1) if self.boss else 1
+        
+        # Clean up boss
+        if self.boss:
+            self.boss.kill()
+        self.boss = None
+        self.boss_active = False
+        
+        # Clear enemies and bullets
+        enemy_count = len(self.enemies)
+        for enemy in self.enemies:
+            enemy.kill()
+        self.enemies.empty()
+        self.enemy_bullets.empty()
+        self.boss_bullets.empty()
+        
+        # Level up
+        old_level = self.game_level
+        is_final_boss = (self.game_level >= MAX_GAME_LEVEL)
+        self.game_level += 1
+        
+        # Score bonus - final boss gets higher bonus
+        if is_final_boss:
+            score_bonus = boss_level * 1000  # Final boss bonus
+        else:
+            score_bonus = boss_level * 500   # Regular boss bonus
+        self.score.update(score_bonus)
+        
+        # Reset timers
+        self.boss_spawn_timer = time.time()
+        self.enemy_spawn_timer = time.time()
+        
+        # Update item spawn interval
+        self.item_spawn_interval = max(15, 30 - self.game_level)
+        
+        # Play sound
+        if hasattr(self, 'sound_manager'):
+            self.sound_manager.play_sound('boss_death.wav')
+        
+        # UI effects
+        self.ui_manager.show_level_up_effects(old_level, self.game_level, enemy_count, score_bonus)
+        
+        # Check for victory
+        if self.game_level > MAX_GAME_LEVEL:
+            self.game_won = True
+            self.ui_manager.show_victory_screen(self.score.value)
+            if hasattr(self, 'sound_manager'):
+                self.sound_manager.fadeout_music(3000)
+    
+    def update_ui_state(self):
+        """
+        Backwards compatibility method for tests.
+        Delegates to the private _update_ui_state method.
+        """
+        self._update_ui_state()
+    
+    def handle_collisions(self, game_time=None):
+        """
+        Backwards compatibility method for tests.
+        Delegates to the private _handle_collisions method.
+        """
+        if game_time is None:
+            game_time = time.time() - self.game_start_time
+        self._handle_collisions(game_time)
+    
+    def spawn_boss(self):
+        """
+        Backwards compatibility method for tests.
+        Delegates to the private _spawn_boss_via_factory method.
+        """
+        return self._spawn_boss_via_factory()
