+++ thunder_fighter/graphics/ui/health_bar.py
